+++ js/modules/base/RObjectPainter.mjs
@@ -2,7 +2,7 @@ import { FontHandler } from './FontHandler.mjs';
 
 import { ObjectPainter } from './ObjectPainter.mjs';
 
-const kNormal = 1, kLessTraffic = 2, kOffline = 3;
+const kNormal = 1, /* kLessTraffic = 2, */ kOffline = 3;
 
 class RObjectPainter extends ObjectPainter {