+++ js/modules/draw/TTree.mjs
@@ -8,7 +8,7 @@ import { kClonesNode, kSTLNode, treeDraw, treeIOTest, TDrawSelector } from '../t
 
 import { BasePainter } from '../base/BasePainter.mjs';
 
-import { cleanup, resize, drawRawText } from '../base/ObjectPainter.mjs';
+import { cleanup, resize, drawRawText, ObjectPainter } from '../base/ObjectPainter.mjs';
 
 import { TH1Painter } from '../hist/TH1Painter.mjs';
 import { TH2Painter } from '../hist/TH2Painter.mjs';
@@ -57,6 +57,8 @@ TDrawSelector.prototype.ShowProgress = function(value) {
    this.last_progress = value;
 }
 
+/** @summary Draw result of tree drawing
+  * @private */
 function drawTreeDrawResult(dom, obj, opt) {
 
    let typ = obj._typename;
@@ -77,6 +79,8 @@ function drawTreeDrawResult(dom, obj, opt) {
 }
 
 
+/** @summary Handle callback function with progress of tree draw
+  * @private */
 function treeDrawProgress(obj, final) {
 
    // no need to update drawing if previous is not yet completed
@@ -239,7 +243,7 @@ function createTreePlayer(player) {
          if (!Number.isInteger(args.firstentry)) delete args.firstentry;
       }
 
-      if (args.drawopt) cleanup(this.drawid);
+      /* if (args.drawopt) */ cleanup(this.drawid);
 
       args.drawid = this.drawid;
 
@@ -381,13 +385,9 @@ function drawLeafPlayer(hpainter, itemname) {
   * @desc just envelope for real TTree::Draw method which do the main job
   * Can be also used for the branch and leaf object
   * @private */
-function drawTree() {
+function drawTree(dom, obj, opt) {
 
-   let painter = this,
-       obj = this.getObject(),
-       opt = this.getDrawOpt(),
-       tree = obj,
-       args = opt;
+   let tree = obj, args = opt;
 
    if (obj._typename == "TBranchFunc") {
       // fictional object, created only in browser
@@ -437,13 +437,16 @@ function drawTree() {
       }
    }
 
+   let painter;
+
    if (args.player) {
+      painter = new ObjectPainter(dom, obj, opt);
       createTreePlayer(painter);
       painter.configureTree(tree);
       painter.showPlayer(args);
       args.drawid = painter.drawid;
    } else {
-      args.drawid = painter.getDom();
+      args.drawid = dom;
    }
 
    // use in result handling same function as for progress handling