+++ js/changes.md
@@ -1,6 +1,24 @@
 # JSROOT changelog
 
 ## Changes in dev
+1. Let change `settings` and `gStyle` parameters via "Settings" menu of the top hierarchy item
+2. Settings and gStyle can be stored as cookies, automatically read with every next load of website
+3. `settings.OnlyLastCycle` defines if only last object version show in TFile (also as `&lastcycle` URL parameter)
+4. `settings.DarkMode` configures dark mode for GUI and drawings (also as `&dark` URL parameter)
+5. Support new `TGraph2DAsymmErrors` class
+6. Support `gStyle.fOptDate` and `gStyle.fOptFile` (also as `&optdate` and `&optfile` URL parameter)
+7. Support `gStyle.fDateX` and `gStyle.fDateY` used for positioning date and file on canvas
+8. Support `gStyle.fHistTopMargin` (also as `&histmargin=value` URL parameter)
+9. Let save frame, title and stats properties to `gStyle` via correspondent context menus
+
+
+## Changes in 7.0.1
+
+1. Fix problem with irregular axis labels
+2. Correctly scale and tilt large number of axes labels
+
+
+## Changes in 7.0.0
 1. Use ES6 modules for code organization
 2. Rewrite code with ES6 classes; one have to use class syntax to derive from it
 3. Upgrade d3.js to 7.3.0, three.js to r138